Add stratified vectorization file
